Fix code and documentation failures
