+++ TriangleNet_Engine/Compute/Representation/GeometricalRepresentation/Element0D/Structure/GeometricalRepresentation.cs
