+++ TriangleNet_Engine/Compute/Representation/GeometricalRepresentation/Element1D/GeometricalRepresentation.cs
