+++ TriangleNet_Engine/Compute/Representation/GeometricalRepresentation/Element1D/Structure/GeometricalRepresentation.cs
