+++ TriangleNet_Engine/Compute/Representation/RenderMesh/Other/RenderMesh.cs
@@ -29,6 +29,7 @@ using System.Linq;
 using System.Text.RegularExpressions;
 using BH.Engine.Geometry;
 using BH.oM.Base;
+using BH.oM.Base.Attributes;
 using System.ComponentModel;
 
 namespace BH.Engine.Representation
@@ -39,26 +40,30 @@ namespace BH.Engine.Representation
         /**** Public Methods - Graphics                 ****/
         /***************************************************/
 
-        public static BH.oM.Graphics.RenderMesh RenderMesh(this BoundingBox bbox, RenderMeshOptions renderMeshOptions = null)
+        [Description("Returns the RenderMesh for the given object, that is a mesh that can be used for Graphical Display.")]
+        [Input("compositeGeometry", "Input composite geometry.")]
+        [Input("renderMeshOptions", "Input renderMeshOptions for how the RenderMesh is computed.")]
+        [Output("renderMesh", "Resulting RenderMesh.")]
+        public static BH.oM.Graphics.RenderMesh RenderMesh(this CompositeGeometry compositeGeometry, RenderMeshOptions renderMeshOptions = null)
         {
-            if (bbox == null)
+            if (compositeGeometry == null)
             {
-                BH.Engine.Base.Compute.RecordError("Cannot compute the mesh of a null bounding box.");
+                BH.Engine.Base.Compute.RecordError("Cannot compute the mesh of a null composite geometry object.");
                 return null;
             }
 
             renderMeshOptions = renderMeshOptions ?? new RenderMeshOptions();
 
-            double length = bbox.Max.X - bbox.Min.X;
-            double depth = bbox.Max.Y - bbox.Min.Y;
-            double height = bbox.Max.Z - bbox.Min.Z;
+            List<RenderMesh> renderMeshes = new List<RenderMesh>();
 
-            Point centrePoint = new Point() { X = bbox.Min.X + length / 2, Y = bbox.Min.Y + depth / 2, Z = bbox.Min.Z + height / 2 };
+            for (int i = 0; i < compositeGeometry.Elements.Count; i++)
+                renderMeshes.Add(IRenderMesh(compositeGeometry.Elements[i]));
 
-            return BoxRenderMesh(centrePoint, length, depth, height);
+            return BH.Engine.Representation.Compute.JoinRenderMeshes(renderMeshes);
         }
     }
 }
+
